Move edit profile sheet out of deprecated profile folder deprecated profile uses edit modal
