--- src/components/profile/index.ts
@@ -1,7 +1,0 @@
-// Profile Exports
-export { BioMessage } from './bio-message';
-export * from './handle-input';
-export { LocationInput } from './location-input';
-export { Profile } from './profile';
-export { ProfileEditModal } from './profile-edit-modal';
-export { ProfileHeader } from './profile-header';
